Include scoping API in secret loader interface
