+++ tracdap-libs/tracdap-lib-common/src/main/java/org/finos/tracdap/common/config/ISecretLoader.java
@@ -60,6 +60,28 @@ public interface ISecretLoader {
     void reload();
 
     /**
+     * Provide a child secret loader restricted to the given scope.
+     *
+     * @param scope The scope for the child secret loader
+     * @return A child secret loader restricted to the given scope
+     */
+    ISecretLoader scope(String scope);
+
+    /**
+     * Convenience method for working with named scopes.
+     *
+     * <p>For example, /tenants/ACME is a named tenant scope
+     *
+     * @param scopeName Name of the scope (a static outer scope)
+     * @param scope Named scope
+     *
+     * @return A child secret loader restricted to the named scope
+     */
+    default ISecretLoader namedScope(String scopeName, String scope) {
+        return scope(scopeName).scope(scope);
+    }
+
+    /**
      * Test whether a secret with the given name exists in the secret store
      *
      * @param secretName The name of the secret to check for
